requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.20.0
pytest==9.0.2
pytest-base-url==2.1.0
pytest-check==2.8.0
pytest-playwright==0.7.2
python-slugify==8.0.4
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
